+++ extensions/browser-bookmarks/src/hooks/useAvailableBrowsers.ts
@@ -23,7 +23,9 @@ export const BROWSERS_BUNDLE_ID = {
   prismaAccess: "com.talon-sec.work",
   vivaldi: "com.vivaldi.vivaldi",
   zen: "app.zen-browser.zen",
+  libreWolf: "org.mozilla.librewolf",
   whale: "com.naver.whale",
+  helium: "net.imput.helium",
 };
 
 export const availableBrowsers = Object.values(BROWSERS_BUNDLE_ID);