+++ extensions/penpot/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Penpot Changelog
 
+## [Windows] - 2025-12-03
+
+- Add Windows support
+- Add Icons for files
+- Group by Teams
+
 ## [Initial Version] - 2024-11-17
 
 Initial version code