+++ extensions/pipedrive/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Pipedrive Changelog
 
+## [Windows Support + Fix CHANGELOG Dates] - 2025-12-04
+
+- Update `shortcut`s to be cross platform
+- Fix Invalid Dates in CHANGELOG
+
 ## [Fix "ccEmail" Copied "title"] - 2025-09-29
 
 - Add `metadata` images
@@ -13,17 +18,17 @@
 
 - Add links to launch Pipedrive to add new person, deal, org or view email or activity when no results are found
 
-## [New Action] - 2025-20-01
+## [New Action] - 2025-01-20
 
 - Added emojis for deal 💰, person 🅿️ and organization 🅾️ to make result types easier to decipher.
 - Updated Pipedrive API from v1 to v2
 - Code cleanup
 
-## [New Action] - 2024-07-10
+## [New Action] - 2024-10-07
 
 - Adds shortcut to copy org name (cmd + o)
 
-## [Bug fix] - 2024-22-01
+## [Bug fix] - 2024-01-22
 
 - Adds ability to search across Pipeline deals, contacts and organizations.
 - Add ability to easily copy search result details (i.e,. name, email, phone number of a contact)