+++ extensions/zipcodebase/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Zipcodebase Changelog
 
+## [Handle Error of invalid response type] - 2025-12-08
+
+- Handle error when invalid response type (ref: [Issue #23499](https://github.com/raycast/extensions/issues/23499))
+
 ## [default_limit in Preferences and state_name in GetPostalCodesByCity] - 2024-03-19
 
 ### Enhancements